% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/markDuplicates.R
\name{bamutilMarkDuplicates}
\alias{bamutilMarkDuplicates}
\title{Mark duplicates using bamutil}
\usage{
bamutilMarkDuplicates(bam, out, path, verbose)
}
\arguments{
\item{bam}{The bam file to mark duplicates from}

\item{out}{Regular expression describing the transformation on the original
filename to get the output filename. By default, a "_duprm" suffix is
added before the bam extension}

\item{path}{Path to the duplicate marker binaries}

\item{verbose}{Redirect all the program output to the R console}
}
\value{
The return code of the system call
}
\description{
\code{bamutilMarkDuplicates} Mark duplicated reads from a BAM file by
  calling bamutil
}
\details{
This function is supposed to be called through the \code{markDuplicates}
  wrapper
}

