% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cumulativeDuprateBarplot.R
\name{cumulativeDuprateBarplot}
\alias{cumulativeDuprateBarplot}
\title{Barplot showing the cumulative read counts fraction}
\usage{
cumulativeDuprateBarplot(DupMat, stepSize = 0.05, ...)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}

\item{stepSize}{The size of the windows used for plotting}

\item{...}{Other params sent to barplot}
}
\value{
nothing
}
\description{
\code{cumulativeDuprateBarplot} Barplot showing the cumulative read counts
  fraction
}
\details{
This function makes a barplot showing the cumulative read counts fraction
from the duplication matrix calculated by \code{analyzeDuprates}.
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# call the plot
cumulativeDuprateBarplot(DupMat=dm)
}

