% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dupRadar.R
\docType{data}
\name{dm.bad}
\alias{dm.bad}
\title{Duplication matrix of a failed RNASeq experiment}
\format{A data frame with 23228 rows and 14 variables}
\usage{
data(dupRadar_examples)
}
\description{
A dataset containing the duplication matrix of a failed RNASeq experiment,
containing unusual duplication rate. Comes from the HCT116 PE2x75 replicate 1
from Caltech (UCSC's table Browser name:
wgEncodeCaltechRnaSeqHct116R2x75Il200AlignsRep1V2)
}
\keyword{datasets}

