% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dupRadar.R
\docType{data}
\name{dm}
\alias{dm}
\title{Duplication matrix of a good RNASeq experiment}
\format{A data frame with 23228 rows and 14 variables}
\usage{
data(dupRadar_examples)
}
\description{
A dataset containing the duplication matrix of a good RNASeq experiment, in
terms of duplicates. Comes from the GM12878 SR1x75 replicate 2 from Caltech
(UCSC's table Browser name: wgEncodeCaltechRnaSeqGm12878R1x75dAlignsRep2V2)
}
\keyword{datasets}

