% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dupRadar.R
\docType{data}
\name{dupRadar_examples}
\alias{dupRadar_examples}
\title{Example data containing precomputed matrices for two RNASeq experiments}
\format{A list with two example duplication matrices}
\usage{
data(dupRadar_examples)
}
\description{
Precomputed duplication matrices for two RNASeq experiments used as examples
of a good and a failed (in terms of high redundancy of reads) experiments.
The experiments come from the ENCODE project, as a source of a broad variety
of protocols, library types and sequencing facilities.
}
\keyword{datasets}

