% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duprateExpBoxplot.R
\name{duprateExpBoxplot}
\alias{duprateExpBoxplot}
\title{Duplication rate ~ total reads per kilobase (RPK) boxplot}
\usage{
duprateExpBoxplot(DupMat, stepSize = 0.05, ...)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}

\item{stepSize}{Expression bin seze for the boxplot}

\item{...}{Other params sent to boxplot}
}
\value{
nothing
}
\description{
\code{duprateExpBoxplot} Duplication rate ~ total reads per kilobase (RPK) boxplot
}
\details{
This function makes a boxplot showing the distribution of per gene duplication rate versus
the reads per kilobase (RPK) inside gene expression bins.
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# duprate boxplot
duprateExpBoxplot(DupMat=dm)
}

