% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duprateExpFit.R
\name{duprateExpFit}
\alias{duprateExpFit}
\title{Duplication rate ~ total read count fit model}
\usage{
duprateExpFit(DupMat)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
The GLM and the coefficients of the fitted logit function
}
\description{
\code{duprateExpDensPlot} Duplication rate ~ total read count fit model
}
\details{
Fit a Generalized Linear Model using a logit function between thegene
duplication rate and the total read count.
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# duprate plot
duprateExpFit(DupMat=dm)
}

