% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expressionHist.R
\name{expressionHist}
\alias{expressionHist}
\title{Draw histogram with the expression values}
\usage{
expressionHist(DupMat, value = "RPK", ...)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}

\item{value}{The column from the duplication matrix containing the
expression values}

\item{...}{Other parameters sent to hist()}
}
\value{
nothing
}
\description{
\code{expressionHist} Draw histogram with the expression values
}
\details{
This function draws a histogram of the expression values from the
duplication matrix calculated by \code{analyzeDuprates}.
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# histogram of expression values for annotation
expressionHist(DupMat=dm)
}

