% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBinDuplication.R
\name{getBinDuplication}
\alias{getBinDuplication}
\title{Helper function used in \code{duprateExpBoxplot}}
\usage{
getBinDuplication(p, stepSize, DupMat)
}
\arguments{
\item{p}{The vector of bins}

\item{stepSize}{The window size}

\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
The duplication rate per bin
}
\description{
\code{getBinDuplication} get duplication rate for a subset of the
duplication matrix
}

