% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBinRpkMean.R
\name{getBinRpkMean}
\alias{getBinRpkMean}
\title{Helper function used in \code{duprateExpBoxplot}}
\usage{
getBinRpkMean(p, stepSize, DupMat)
}
\arguments{
\item{p}{The vector of bins}

\item{stepSize}{The window size}

\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
The averaged RPK per bin
}
\description{
\code{getBinRpkMean} get mean duplication rate per bin
}

