% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDupMatBin.R
\name{getDupMatBin}
\alias{getDupMatBin}
\title{Helper function used in \code{getBinDuplication} and
  \code{getBinRpkMean}}
\usage{
getDupMatBin(p, stepSize = 0.05, value = "allCounts", DupMat)
}
\arguments{
\item{p}{The vector of bins}

\item{stepSize}{The window size}

\item{value}{The column to be subset}

\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
The subseted matrix
}
\description{
\code{getDupMatBin} get a subset of the matrix for values in a specific bin
defined by the upper bound p and stepSize
}

