% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDupMatStats.R
\name{getDupMatStats}
\alias{getDupMatStats}
\title{Report duplication stats on regions}
\usage{
getDupMatStats(DupMat)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
A data.frame containing the stats about the number of genes covered
  (1+ tags) and the number of genes containing duplicates (1+)
}
\description{
\code{getDupMatStats} Report duplication stats based on the data calculated
  in the duplication matrix
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# call the plot and identify genes
getDupMatStats(DupMat=dm)
}

