% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDynamicRange.R
\name{getDynamicRange}
\alias{getDynamicRange}
\title{Dynamic range}
\usage{
getDynamicRange(dm)
}
\arguments{
\item{dm}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
A list with 2 elements, containing the dynamic range counting all
reads and the dynamic range after removing duplicates.
}
\description{
\code{getDynamicRange} Calculate the dynamic range of the RNAseq experiment
}
\details{
This function calculates the dynamic range of the RNAseq eperiment
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)

# calculate the dynamic range
getDynamicRange(dm)
}

