% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRpkBinReadCountFraction.R
\name{getRpkBinReadCountFraction}
\alias{getRpkBinReadCountFraction}
\title{Helper function used in \code{readcountExpressionBoxplot}}
\usage{
getRpkBinReadCountFraction(p, stepSize = stepSize, DupMat = DupMat)
}
\arguments{
\item{p}{The vector of bins}

\item{stepSize}{The window size}

\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}
}
\value{
The fraction of total reads in a vector of bins
}
\description{
\code{readcountExpressionBoxplot} Calculates the fraction of total reads in
  a vector of bins
}

