% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readcountExpBoxplot.R
\name{readcountExpBoxplot}
\alias{readcountExpBoxplot}
\title{Barplot of percentage of reads falling into expression bins}
\usage{
readcountExpBoxplot(DupMat, stepSize = 0.05, ...)
}
\arguments{
\item{DupMat}{The duplication matrix calculated by \code{analyzeDuprates}}

\item{stepSize}{The number of bars to be shown}

\item{...}{Other parameters sent to barplot()}
}
\value{
nothing Other parameters sent to barplot()
}
\description{
\code{readcountExpBoxplot} Barplot of percentage of reads falling into
expression bins
}
\details{
This function makes a barplot of percentage of reads falling into expression
bins
}
\examples{
# dm is a duplication matrix calculated by analyzeDuprates:
# R> dm <- analyzeDuprates(bamDuprm,gtf,stranded,paired,threads)
attach(dupRadar_examples)


# barplot of percentage of reads falling into expression bins
readcountExpBoxplot(DupMat=dm)
}

